+++ lib/Calibre/Data.php
@@ -279,6 +279,9 @@ class Data
      */
     public static function getLink($book, $type, $mime, $rel, $filename, $idData, $title = null, $view = false)
     {
+        if (!empty(Config::get('calibre_external_storage')) && str_starts_with($book->path, Config::get('calibre_external_storage'))) {
+            return new LinkEntry($book->path . "/" . rawurlencode($filename), $mime, $rel, $title);
+        }
         // moved image-specific code from Data to Cover
         if (Database::useAbsolutePath($book->getDatabaseId()) ||
             ($type == "epub" && Config::get('update_epub-metadata'))) {