+++ CHANGELOG.md
@@ -11,7 +11,10 @@ x.x.x - TODO
   * ...
 
 2.7.x - 2024xxxx
+  * Changes in config_default.php file:
+    - new $config['calibre_external_storage']
   * ...
+  * Support external storage for Calibre library - see seblucas/cops#506 and seblucas/cops#513
   * Pass along request handler in baselist, booklist and virtual libraries
   * Adjust default settings for epubjs-reader - see pull request #81 from @dunxd
   * Rename IndexHandler to HtmlHandler and use default 'index' in request